TcpChannel -> SymChannel, tidy
