+++ fancy-garbling/src/util.rs
@@ -12,7 +12,7 @@ use crate::Wire;
 #[cfg(feature = "nightly")]
 use core::arch::x86_64::*;
 use itertools::Itertools;
-use scuttlebutt::{Block, Block512};
+use scuttlebutt::Block;
 use std::collections::HashMap;
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -372,43 +372,6 @@ pub fn generate_deltas(primes: &[u16]) -> HashMap<u16, Wire> {
         deltas.insert(*q, Wire::rand_delta(&mut rng, *q));
     }
     deltas
-}
-
-//Assumes payloads are up to 64bit long i.e 8 bytes
-fn block512_to_crt(b: Block512) -> Vec<u16> {
-    let b_val = b.prefix(8);
-
-    let mut b_128 = [0_u8; 16];
-    b_128[..8].clone_from_slice(&b_val[..8]);
-
-    let q = primes_with_width(64);
-    crt(u128::from_le_bytes(b_128), &q)
-}
-
-//Assumes payloads are up to 64bit long
-// WRITE assumption more
-pub fn mask_payload_crt<RNG: rand::Rng + Sized>(
-    x: Block512,
-    y: Block512,
-    rng: &mut RNG,
-) -> Block512 {
-    let x_crt = block512_to_crt(x);
-    let y_crt = block512_to_crt(y);
-    let q = primes_with_width(64);
-    let mut res_crt = Vec::new();
-    for i in 0..q.len() {
-        res_crt.push((x_crt[i] + y_crt[i]) % q[i]);
-    }
-    let res = crt_inv(&res_crt, &q).to_le_bytes();
-    let mut block = [0_u8; 64];
-    for i in 0..64 {
-        if i < res.len() {
-            block[i] = res[i];
-        } else {
-            block[i] = rng.gen::<u8>(); // TODO: mod rest of prime
-        }
-    }
-    Block512::from(block)
 }
 
 /// Extra Rng functionality, useful for `fancy-garbling`.
