+++ popsicle/src/lib.rs
@@ -16,4 +16,4 @@ mod errors;
 mod psi;
 pub mod utils;
 
-pub use crate::{errors::Error, psi::*, utils::parse_files};
+pub use crate::{errors::Error, psi::*};
