+++ popsicle/src/utils.rs
@@ -5,12 +5,8 @@
 // See LICENSE for licensing information.
 
 use rand::{CryptoRng, Rng};
-use scuttlebutt::{AesHash, Block, Block512};
+use scuttlebutt::{AesHash, Block};
 use sha2::{Digest, Sha256};
-use std::{
-    fs::File,
-    io::{BufRead, BufReader},
-};
 
 // Compress an arbitrary vector into a 128-bit chunk, leaving the final 8-bits
 // as zero. We need to leave 8 bits free in order to add in the hash index when
@@ -37,54 +33,6 @@ pub fn compress_and_hash_inputs(inputs: &[Vec<u8>], key: Block) -> Vec<Block> {
             block & mask
         })
         .collect::<Vec<Block>>()
-}
-
-pub fn int_vec_block512(values: Vec<u64>) -> Vec<Block512> {
-    values
-        .into_iter()
-        .map(|item| {
-            let value_bytes = item.to_le_bytes();
-            let mut res_block = [0_u8; 64];
-            res_block.clone_from_slice(&value_bytes[..8]);
-            Block512::from(res_block)
-        })
-        .collect()
-}
-
-pub fn parse_files(
-    id_position: usize,
-    payload_position: usize,
-    path: &str,
-) -> (Vec<Vec<u8>>, Vec<Block512>) {
-    let data = File::open(path).unwrap();
-
-    let buffer = BufReader::new(data).lines();
-
-    let mut ids = Vec::new();
-    let mut payloads = Vec::new();
-
-    let mut cnt = 0;
-    for line in buffer.enumerate() {
-        let line_split = line
-            .1
-            .unwrap()
-            .split(',')
-            .map(|item| item.to_string())
-            .collect::<Vec<String>>();
-        if cnt == 0 {
-            cnt += 1;
-        } else {
-            ids.push(
-                line_split[id_position]
-                    .parse::<u64>()
-                    .unwrap()
-                    .to_le_bytes()
-                    .to_vec(),
-            );
-            payloads.push(line_split[payload_position].parse::<u64>().unwrap());
-        }
-    }
-    (ids, int_vec_block512(payloads))
 }
 
 #[allow(dead_code)] // used in tests
